Changed param in lesson method and added mark method for getting mark with lesson or task entity

+++ EStudy/EStudy.Domain/Interfaces/ILessonRepository.cs
@@ -8,6 +8,6 @@ namespace EStudy.Domain.Interfaces
 {
     public interface ILessonRepository : IRepository<Lesson>
     {
-        Task<Lesson> GetLessonWithSubjectById(int Id);
+        Task<Lesson> GetLessonWithSubjectById(long Id);
     }
 }
